rename/#9 : ProfileViewHolder 이름 변경
